Update VB side.
